Reorganized file structure and fixed imports. Ensured consistent naming conventions.
