+++ Red River Mart/src/App.tsx
@@ -1,7 +1,7 @@
 import "./App.css";
-import Header from "./headerComponent/header";
-import LandingPageMain from './components/LandingPageMainComponent/LandingPageMain'
-import Footer from './components/footer'
+import Header from "./components/headerComponent/header";
+import LandingPageMain from './components/buyComponent/buyPage'
+import Footer from './components/footerComponent/Footer';
 
 function App() {
   return (
