+++ Red River Mart/src/components/footerComponent/Footer.tsx
